Add 'scaleUp' option
This option lets client select whether to generate a thumbnail when it
would scale up the image instead.

It is enabled by default (current behaviour), if client wants to
generate
a thumbnail only unless the image is going to scale up, just set
'scaleUp'
to false and the image size won't change.

See #87, #85